feat(ADS-7): consolidate preference enums into lib.types shared package
Move ProfileVisibility and DigestFrequency enums from their respective
backend model files (UserPrivacyPrefs, UserNotificationPrefs) into
lib.types so backend models and frontend forms share a single source of
truth for these types.

- Add lib.types/src/types/preferences.ts with ProfileVisibility and
  DigestFrequency enums
- Backend models now import from @adopt-dont-shop/lib.types and re-export
  for backward compatibility with existing service imports
- Add lib.types as a direct dependency in app.client
- Update User.privacySettings.profileVisibility type from `string` to
  ProfileVisibility in app.client types
- Update SettingsForm and ProfilePage to use ProfileVisibility enum;
  adds the previously missing `rescues_only` dropdown option
- Fix lib.types tsconfig.cjs.json ignoreDeprecations for TypeScript 6.x

https://linear.app/ideasquared/issue/ADS-7
