+++ lib.types/tsconfig.cjs.json
@@ -4,7 +4,7 @@
     "target": "ES2020",
     "rootDir": "./src",
     "outDir": "./dist/cjs",
-    "ignoreDeprecations": "5.0",
+    "ignoreDeprecations": "6.0",
     "declaration": false,
     "declarationMap": false,
     "sourceMap": false,
